requirements.txt
albumentations==1.0.0
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
cycler==0.10.0
decorator==4.4.2
Django==3.2.4
django-sslserver
einops
efficientnet-pytorch==0.7.1
idna==2.10
imageio==2.9.0
kiwisolver==1.3.1
librosa
matplotlib
networkx==2.5.1
numpy==1.22.0
opencv-python-headless
Pillow
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
requests
scikit-image==0.18.1
scipy
setuptools
six==1.16.0
sqlparse==0.4.1
tifffile==2021.6.14
-f https://download.pytorch.org/whl/torch_stable.html
torch==1.9.0+cpu
torchaudio
torchvision==0.10.0+cpu
typing-extensions==3.10.0.0
urllib3
Blockchain
Block


